add to cart

+++ controllers/product.js
@@ -61,14 +61,18 @@ exports.read = async (req, res) => {
 exports.update = async (req, res) => {
     const productId = req.params.productId;
     
-    req.body.fileName = req.file.filename;
+    if(req.file !== undefined){
+        req.body.fileName = req.file.filename;
+    }
 
     const oldProduct = await Product.findByIdAndUpdate(productId, req.body);
 
-    fs.unlink(`uploads/${oldProduct.fileName}`, err => {
-        if (err) throw err;
-        console.log("old image successfully deleted.........");
-    });
+    if(req.file !== undefined && req.file.filename !== oldProduct.fileName){
+        fs.unlink(`uploads/${oldProduct.fileName}`, err => {
+            if (err) throw err;
+            console.log("old image successfully deleted.........");
+        });
+    }
 
     res.json({
         successMessage: 'Product successfully updated...'
